Separate User/Group functions into zosaccount file
