+++ c/zosfile.c
@@ -825,111 +825,6 @@ int fileChangeOwner(const char *fileName, int *returnCode, int *reasonCode,
   return returnValue;
 }
 
-int gidGetUserInfo(const char *userName,  UserInfo * info,
-                         int *returnCode, int *reasonCode) {
-  int nameLength = strlen(userName);
-  int *reasonCodePtr;
-  int returnValue;
-  int retValue = -1;
-
-  UserInfo *ptrInfo;
-
-#ifndef _LP64
-  reasonCodePtr = (int*) (0x80000000 | ((int)reasonCode));
-#else
-  reasonCodePtr = reasonCode;
-#endif
-
-  BPXGPN(&nameLength,
-         userName,
-         &returnValue,
-         returnCode,
-         reasonCodePtr);
-
-  /* Copy returned structure */
-  if (returnValue != 0) {
-    memcpy (info, (char *)returnValue, sizeof (UserInfo));
-    retValue = 0;
-  }
-
-  if (fileTrace) {
-    if(returnValue == 0) {
-#ifdef METTLE
-      printf("BPXGPN (%s) FAILED: returnValue: %d, returnCode: %d, reasonCode: 0x%08x\n",
-             userName, returnValue, *returnCode, *reasonCode);
-#else
-      printf("BPXGPN (%s) FAILED: returnValue: %d, returnCode: %d, reasonCode: 0x%08x, strError: (%s)\n",
-             userName, returnValue, *returnCode, *reasonCode, strerror(*returnCode));
-#endif
-    }
-    else {
-      printf("BPXGPN (%s) OK: returnVal: %d\n", userName, returnValue);
-    }
-  }
-
-  return retValue;
-}
-
-int gidGetGroupInfo(const char *groupName,  GroupInfo *info,
-                   int *returnCode, int *reasonCode) {
-  int groupLength = strlen(groupName);
-  int *reasonCodePtr;
-  int returnValue;
-  int retValue = -1;
-  *returnCode = *reasonCode = 0;
-
-#ifndef _LP64
-  reasonCodePtr = (int*) (0x80000000 | ((int)reasonCode));
-#else
-  reasonCodePtr = reasonCode;
-#endif
-
-  BPXGGN(&groupLength,
-         groupName,
-         &returnValue,
-         returnCode,
-         reasonCodePtr);
-
-  /* Copy returned structure */
-  if (returnValue >  0) {
-    memcpy (info, (char *)returnValue, sizeof (GroupInfo));
-    retValue = 0;
-  }
-
-  if (fileTrace) {
-    if(returnValue == 0) {
-#ifdef METTLE
-      printf("BPXGGN (%s) FAILED: returnValue: %d, returnCode: %d, reasonCode: 0x%08x\n",
-             groupName, returnValue, *returnCode, *reasonCode);
-#else
-      printf("BPXGGN (%s) FAILED: returnValue: %d, returnCode: %d, reasonCode: 0x%08x, strError: (%s)\n",
-             groupName, returnValue, *returnCode, *reasonCode, strerror(*returnCode));
-#endif
-    }
-    else {
-      printf("BPXGGN (%s) OK: returnVal: %d\n", groupName, returnValue);
-    }
-  }
-
-  return retValue;
-}
-
-int userInfoGetUserId (UserInfo *info) {
-  int *temp = (int *)info;
-  int unameLength = info->GIDN_U_LEN;
-  int unameLengthindex = (unameLength + 3) /4;
-  int userId = temp[unameLengthindex+ 2];
-  return userId;
-}
-
-int groupInfoGetGroupId (GroupInfo *info) {
-  int *temp = (int *)info;
-  int groupLength = info->GIDS_G_LEN;
-  int groupLengthindex = (groupLength + 3) /4;
-  int groupId = temp[groupLengthindex+ 2];
-  return groupId;
-}
-
 int fileInfoIsDirectory(const FileInfo *info) {
   return (info->fileType == BPXSTA_FILETYPE_DIRECTORY ? TRUE: FALSE);
 }
